jwxt_course/requirements.txt
absl-py==0.12.0
asgiref==3.3.1
astroid==2.3.3
astunparse==1.6.3
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.5
cachetools==4.2.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
Django==3.1.5
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.8.2
flatbuffers==1.12
fr-utils==0.0.1
gast==0.3.3
gdown==4.6.0
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.8
imageio==2.9.0
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.0
jieba==0.42.1
Jinja2==2.11.2
joblib==1.0.1
jsonpatch==1.32
jsonpointer==2.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
loguru==0.6.0
lxml @ file:///C:/Users/%E8%81%94%E6%83%B3/Desktop/%E5%A4%87%E4%BB%BD/python/lxml-4.7.1-cp38-cp38-win_amd64.whl
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
mccabe==0.6.1
mistune==0.8.4
MultiScaleDeformableAttention==1.0
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.5.1
notebook==6.0.3
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.2.54
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==20.4
pandas==1.2.4
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyExecJS2==1.6.1
pygame==1.9.6
Pygments==2.6.1
pylint==2.4.4
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.3
seaborn==0.11.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
snownlp==0.12.3
soupsieve==1.9.5
sqlparse==0.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.8.3
tesserocr @ file:///C:/Users/%E8%81%94%E6%83%B3/Desktop/%E5%A4%87%E4%BB%BD/python/tesserocr-2.5.2-cp38-cp38-win_amd64.whl
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.6.6
torch==1.4.0+cu92
torchfile==0.1.0
torchnet==0.0.4
torchvision==0.5.0
tornado==6.0.4
tqdm==4.64.1
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.7
visdom==0.1.8.9
wcwidth==0.2.3
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win32-setctime==1.1.0
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
